カラムをfloat型に変更

+++ db/migrate/20210112025054_create_calculations.rb
@@ -1,8 +1,8 @@
 class CreateCalculations < ActiveRecord::Migration[6.0]
   def change
     create_table :calculations do |t|
-      t.integer :weight, null: false
-      t.integer :height, null: false
+      t.float :weight, null: false
+      t.float :height, null: false
       t.timestamps
     end
   end
