+++ db/schema.rb
@@ -13,8 +13,8 @@
 ActiveRecord::Schema.define(version: 2021_01_12_025054) do
 
   create_table "calculations", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
-    t.integer "weight", null: false
-    t.integer "height", null: false
+    t.float "weight", null: false
+    t.float "height", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
   end
